+++ src/styleSheets/components/_Home.scss
@@ -1,23 +1,19 @@
 @import url("https://fonts.googleapis.com/css2?family=Lovers+Quarrel&display=swap");
 $font-title: "Lovers Quarrel", cursive;
 
-// ? background image mixin
+// !---MIXINS_&_VARIABLES--------------------------------------------------------------
 @mixin background-image($image) {
   background-image: url($image);
   background-size: cover;
   background-position: center;
   background-repeat: no-repeat;
 }
-
-// ? flex mixin
 @mixin flex($justify, $align, $direction) {
   display: flex;
   justify-content: $justify;
   align-items: $align;
   flex-direction: $direction;
 }
-
-// ? Gradient mixin
 $start-color: #2d2d2d;
 $end-color: #000000;
 @mixin radial-gradient-dark($start-color, $end-color) {
@@ -30,12 +26,11 @@ body {
   font-size: $font-size-3;
 }
 
-// ! FIRST 100VH SECTION ------------------------------------------------------------------
+// !---Home_Header.jsx------------------------------------------------------------------
 #background {
   @include background-image("src/assets/images/castles/iceTower.jpg");
   height: 100vh;
 
-  // ! HEADER component
   #home-header {
     @include flex(flex-end, center, row);
     background-color: rgba($color: #000000, $alpha: 0.5);
@@ -59,7 +54,7 @@ body {
     }
   }
 
-  // ! HERO ONE component
+  // !---Home_HeroOne.jsx---------------------------------------------------------------
   #heroOne {
     @include flex(center, center, row);
     height: 85vh;
@@ -87,7 +82,6 @@ body {
         color: rgb(1, 0, 0);
         border: none;
         padding: 1rem 2rem;
-        /* font-size: 1.5rem; */
         border-radius: $border-radius-4;
         cursor: pointer;
         box-shadow: 0 0 10px rgba($color: #ffffff, $alpha: 0.7);
@@ -96,43 +90,18 @@ body {
   }
 }
 
-// ! SECOND 100VH SECTION ------------------------------------------------------------------
-#hero-divider {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
+// !---Home_HeroTwo.jsx---------------------------------------------------------------
+#heroTwo {
+  @include flex(center, center, row);
+  @include radial-gradient-dark($start-color, $end-color);
   height: 100vh;
-  background: radial-gradient(
-    ellipse at center,
-    #1c1c1c 0%,
-    #000000 100%
-  ); // Radial gradient background
 
-  #divider-left,
-  #divider-right {
+  #heroTwo-left,
+  #heroTwo-right {
+    @include flex(center, center, column);
     width: 50%;
     height: 100%;
     padding: 6rem;
-
-    h2 {
-      font-family: $font-title;
-      font-size: 5rem;
-      color: #fff;
-      margin-bottom: 2rem;
-    }
-
-    p {
-      font-size: 1.5rem;
-      color: #fff;
-      line-height: 1.5;
-    }
-  }
-
-  #divider-left {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
 
     img {
       background-color: rgba($color: #000000, $alpha: 0.6);
@@ -142,42 +111,38 @@ body {
       width: 80%;
       height: auto;
     }
-  }
 
-  #divider-right {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: flex-start;
+    h2 {
+      font-family: $font-title;
+      font-size: 5rem;
+      margin-bottom: 2rem;
+    }
+
+    p {
+      line-height: 1.5;
+    }
   }
 }
 
-// ! THIRD 100VH SECTION ------------------------------------------------------------------
-#playingCardsHero {
-  display: flex;
-  justify-content: center;
-  align-items: center;
+// !---Home_HeroThree.jsx------------------------------------------------------------------
+#heroThree {
+  @include flex(flex-start, center, row);
+  @include background-image("src/assets/images/landscapes/background-10.jpg");
   height: 100vh;
-  background-image: url("/src/assets/images/landscapes/background-11.jpg");
-  background-position: center center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  color: white;
-  #playingCards {
-    width: 100%;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    text-align: center;
 
+  #playingCards {
+    @include flex(flex-start, center, row);
+    justify-content: space-evenly;
+    text-align: center;
+    height: 100%;
+    overflow: auto;
     .card-big {
+      @include flex(center, center, column);
       background-color: rgba($color: #000000, $alpha: 0.6);
       box-shadow: 0 0 10px rgba($color: #ffffff, $alpha: 0.7);
       border-radius: $border-radius-4;
-      width: 200px;
+      width: 250px;
       height: auto;
-      margin: 2rem;
 
       h2 {
         font-family: $font-title;
